+++ src/storage/index.ts
@@ -1,3 +1,5 @@
 export * as StorageUtils from './StorageUtils.js'
 
 export { StorageUploader } from './StorageUploader.js'
+
+export { StorageDownloader } from './StorageDownloader.js'