+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file. The format 
 ## Table of Contents
 
 - [Unreleased](#unreleased)
-- [1.3.36 - 2025-03-07](#1335---2025-03-7)
+- [1.4.2 - 2025-03-13](#142---2025-03-13)
+- [1.4.1 - 2025-03-12](#141---2025-03-12)
+- [1.4.0 - 2025-03-10](#140---2025-03-7)
+  - Scope increase to include Storage and Message Box client functionality
+- [1.3.36 - 2025-03-07](#1336---2025-03-7)
 - [1.3.35 - 2025-03-07](#1335---2025-03-7)
 - [1.3.34 - 2025-03-06](#1334---2025-03-6)
 - [1.3.33 - 2025-03-06](#1333---2025-03-6)
@@ -100,6 +104,34 @@ All notable changes to this project will be documented in this file. The format 
 ### Security
 
 ---
+
+## [1.4.2] - 2025-03-13
+
+### Added
+
+- Implement UHRP Storage Downloader
+
+---
+
+## [1.4.1] - 2025-03-12
+
+### Added
+
+- Implement IdentityClient
+- Implement RegistryClient
+
+---
+
+## [1.4.0] - 2025-03-10
+
+### Added
+
+- Implement UHRP Storage Uploader
+- Implement UHRP Storage Downloader
+- Test coverage WIP
+
+---
+
 
 ## [1.3.36] - 2025-03-7